backend/requirements.txt
torch
torchvision
diffusers
transformers
fastapi
uvicorn
numpy
pillow
tqdm
scipy
trimesh
scikit-image 